Separate gateway

+++ jobsdb/jobsdb.go
@@ -160,6 +160,7 @@ type HandleT struct {
 	migrationState                MigrationState
 	inProgressMigrationTargetDS   *dataSetT
 	logger                        logger.LoggerI
+	ownerType                     OwnerType
 }
 
 //The struct which is written to the journal
@@ -226,7 +227,7 @@ func (jd *HandleT) assert(cond bool, errorString string) {
 	if !cond {
 		jd.printLists(true)
 		jd.logger.Fatal(jd.dsEmptyResultCache)
-		panic(errorString)
+		panic(fmt.Errorf("[[ %s ]]: %s", jd.tablePrefix, errorString))
 	}
 }
 
@@ -289,6 +290,18 @@ var jobStates []jobStateT = []jobStateT{
 	WontMigrate,
 }
 
+//OwnerType for this jobsdb instance
+type OwnerType string
+
+const (
+	//Read : Only Reader of this jobsdb instance
+	Read OwnerType = "READ"
+	//Write : Only Writer of this jobsdb instance
+	Write OwnerType = "WRITE"
+	//ReadWrite : Reader and Writer of this jobsdb instance
+	ReadWrite OwnerType = ""
+)
+
 func getValidStates() (validStates []string) {
 	for _, js := range jobStates {
 		if js.isValid {
@@ -339,6 +352,7 @@ var (
 	jobDoneMigrateThres, jobStatusMigrateThres   float64
 	migrateDSLoopSleepDuration                   time.Duration
 	addNewDSLoopSleepDuration                    time.Duration
+	refreshDSListLoopSleepDuration               time.Duration
 	backupCheckSleepDuration                     time.Duration
 	useJoinForUnprocessed                        bool
 	backupRowsBatchSize                          int64
@@ -370,6 +384,7 @@ func loadConfig() {
 	maxMigrateDSProbe: Maximum number of DSs that are checked from left to right if they are eligible for migration
 	migrateDSLoopSleepDuration: How often is the loop (which checks for migrating DS) run
 	addNewDSLoopSleepDuration: How often is the loop (which checks for adding new DS) run
+	refreshDSListLoopSleepDuration: How often is the loop (which refreshes DSList) run
 	maxTableSizeInMB: Maximum Table size in MB
 	*/
 	jobDoneMigrateThres = config.GetFloat64("JobsDB.jobDoneMigrateThres", 0.8)
@@ -381,6 +396,7 @@ func loadConfig() {
 	backupRowsBatchSize = config.GetInt64("JobsDB.backupRowsBatchSize", 1000)
 	migrateDSLoopSleepDuration = (config.GetDuration("JobsDB.migrateDSLoopSleepDurationInS", time.Duration(30)) * time.Second)
 	addNewDSLoopSleepDuration = (config.GetDuration("JobsDB.addNewDSLoopSleepDurationInS", time.Duration(5)) * time.Second)
+	refreshDSListLoopSleepDuration = (config.GetDuration("JobsDB.refreshDSListLoopSleepDurationInS", time.Duration(5)) * time.Second)
 	backupCheckSleepDuration = (config.GetDuration("JobsDB.backupCheckSleepDurationIns", time.Duration(2)) * time.Second)
 	useJoinForUnprocessed = config.GetBool("JobsDB.useJoinForUnprocessed", true)
 
@@ -406,8 +422,9 @@ multiple users of JobsDB
 dsRetentionPeriod = A DS is not deleted if it has some activity
 in the retention time
 */
-func (jd *HandleT) Setup(clearAll bool, tablePrefix string, retentionPeriod time.Duration, migrationMode string, registerStatusHandler bool) {
+func (jd *HandleT) Setup(ownerType OwnerType, clearAll bool, tablePrefix string, retentionPeriod time.Duration, migrationMode string, registerStatusHandler bool) {
 
+	jd.ownerType = ownerType
 	jd.logger = pkgLogger.Child(tablePrefix)
 	var err error
 	jd.migrationState.migrationMode = migrationMode
@@ -422,9 +439,6 @@ func (jd *HandleT) Setup(clearAll bool, tablePrefix string, retentionPeriod time
 
 	jd.BackupSettings = jd.getBackUpSettings()
 
-	//Kill any pending queries
-	jd.terminateQueries()
-
 	jd.dbHandle, err = sql.Open("postgres", psqlInfo)
 	jd.assertError(err)
 
@@ -437,9 +451,62 @@ func (jd *HandleT) Setup(clearAll bool, tablePrefix string, retentionPeriod time
 	jd.statNewDSPeriod = stats.NewStat(fmt.Sprintf("jobsdb.%s_new_ds_period", jd.tablePrefix), stats.TimerType)
 	jd.statDropDSPeriod = stats.NewStat(fmt.Sprintf("jobsdb.%s_drop_ds_period", jd.tablePrefix), stats.TimerType)
 
-	jd.setupDatabaseTables(clearAll)
+	switch ownerType {
+	case Read:
+		jd.readerSetup()
+	case Write:
+		jd.setupDatabaseTables(clearAll)
+		jd.writerSetup()
+	case ReadWrite:
+		jd.setupDatabaseTables(clearAll)
+		jd.readerWriterSetup()
+	}
+}
 
-	jd.recoverFromJournal()
+func (jd *HandleT) startBackupDSLoop() {
+	var err error
+	if jd.BackupSettings.BackupEnabled {
+		jd.jobsFileUploader, err = jd.getFileUploader()
+		jd.assertError(err)
+		rruntime.Go(func() {
+			jd.backupDSLoop()
+		})
+	}
+
+}
+
+func (jd *HandleT) startMigrateDSLoop() {
+	rruntime.Go(func() {
+		jd.migrateDSLoop()
+	})
+}
+
+func (jd *HandleT) readerSetup() {
+	jd.recoverFromJournal(Read)
+
+	//This is a thread-safe operation.
+	//Even if two different services (gateway and processor) perform this operation, there should not be any problem.
+	jd.recoverFromJournal(ReadWrite)
+
+	//Refresh in memory list. We don't take lock
+	//here because this is called before anything
+	//else
+	jd.getDSList(true)
+	jd.getDSRangeList(true)
+
+	rruntime.Go(func() {
+		jd.refreshDSListLoop()
+	})
+
+	jd.startBackupDSLoop()
+	jd.startMigrateDSLoop()
+}
+
+func (jd *HandleT) writerSetup() {
+	jd.recoverFromJournal(Write)
+	//This is a thread-safe operation.
+	//Even if two different services (gateway and processor) perform this operation, there should not be any problem.
+	jd.recoverFromJournal(ReadWrite)
 
 	//Refresh in memory list. We don't take lock
 	//here because this is called before anything
@@ -452,20 +519,18 @@ func (jd *HandleT) Setup(clearAll bool, tablePrefix string, retentionPeriod time
 		jd.addNewDS(appendToDsList, dataSetT{})
 	}
 
-	if jd.BackupSettings.BackupEnabled {
-		jd.jobsFileUploader, err = jd.getFileUploader()
-		jd.assertError(err)
-		rruntime.Go(func() {
-			jd.backupDSLoop()
-		})
-	}
-	rruntime.Go(func() {
-		jd.migrateDSLoop()
-	})
-
 	rruntime.Go(func() {
 		jd.addNewDSLoop()
 	})
+}
+
+func (jd *HandleT) readerWriterSetup() {
+	jd.recoverFromJournal(Read)
+
+	jd.writerSetup()
+
+	jd.startBackupDSLoop()
+	jd.startMigrateDSLoop()
 }
 
 /*
@@ -560,6 +625,28 @@ func (jd *HandleT) getAllTableNames() []string {
 	return tableNames
 }
 
+//removeExtraKey : removes extra key present in map1 and not in map2
+//Assumption is keys in map1 and map2 are same, except that map1 has one key more than map2
+func removeExtraKey(map1, map2 map[string]string) string {
+	var deleteKey, key string
+	for key = range map1 {
+		if _, ok := map2[key]; !ok {
+			deleteKey = key
+			break
+		}
+	}
+
+	if deleteKey != "" {
+		delete(map1, deleteKey)
+	}
+
+	return deleteKey
+}
+
+func remove(slice []string, idx int) []string {
+	return append(slice[:idx], slice[idx+1:]...)
+}
+
 /*
 Function to return an ordered list of datasets and datasetRanges
 Most callers use the in-memory list of dataset and datasetRanges
@@ -602,11 +689,26 @@ func (jd *HandleT) getDSList(refreshFromDB bool) []dataSetT {
 		}
 	}
 
-	if len(dnumList) == 0 {
-		return jd.datasetList
-	}
-
 	jd.sortDnumList(dnumList)
+
+	//If any service has crashed while creating DS, this may happen. Handling such case gracefully.
+	if len(jobNameMap) != len(jobStatusNameMap) {
+		jd.assert(len(jobNameMap) == len(jobStatusNameMap)+1, fmt.Sprintf("Length of jobNameMap(%d) - length of jobStatusNameMap(%d) is more than 1", len(jobNameMap), len(jobStatusNameMap)))
+		deletedDNum := removeExtraKey(jobNameMap, jobStatusNameMap)
+		//remove deletedDNum from dnumList
+		var idx int
+		var dnum string
+		var foundDeletedDNum bool
+		for idx, dnum = range dnumList {
+			if dnum == deletedDNum {
+				foundDeletedDNum = true
+				break
+			}
+		}
+		if foundDeletedDNum {
+			dnumList = remove(dnumList, idx)
+		}
+	}
 
 	//Create the structure
 	for _, dnum := range dnumList {
@@ -618,6 +720,16 @@ func (jd *HandleT) getDSList(refreshFromDB bool) []dataSetT {
 			dataSetT{JobTable: jobName,
 				JobStatusTable: jobStatusName, Index: dnum})
 	}
+
+	//if the owner of this jobsdb is a writer, then shrinking datasetList to have only last two datasets
+	//this shrinked datasetList is used to compute DSRangeList
+	//This is done because, writers don't care about the left datasets in the sorted datasetList
+	if jd.ownerType == Write {
+		if len(jd.datasetList) > 2 {
+			jd.datasetList = jd.datasetList[len(jd.datasetList)-2 : len(jd.datasetList)]
+		}
+	}
+
 	jd.statTableCount.Gauge(len(jd.datasetList))
 	return jd.datasetList
 }
@@ -800,6 +912,20 @@ func (jd *HandleT) createTableNames(dsIdx string) (string, string) {
 	return jobTable, jobStatusTable
 }
 
+func (jd *HandleT) addNewDSonSetup(idx string) dataSetT {
+	queryStat := stats.NewTaggedStat("add_new_ds", stats.TimerType, stats.Tags{"customVal": jd.tablePrefix})
+
+	queryStat.Start()
+	defer queryStat.End()
+
+	ds := jd.createDS(true, idx)
+
+	jd.statNewDSPeriod.Start()
+	jd.isStatNewDSPeriodInitialized = true
+
+	return ds
+}
+
 func (jd *HandleT) addNewDS(newDSType string, insertBeforeDS dataSetT) dataSetT {
 	jd.logger.Infof("Creating new DS of type %s before ds %s for %s jobsdb", newDSType, insertBeforeDS.Index, jd.tablePrefix)
 	var newDSIdx string
@@ -821,7 +947,7 @@ func (jd *HandleT) addNewDS(newDSType string, insertBeforeDS dataSetT) dataSetT 
 		panic("Unknown usage for newDSType : " + newDSType)
 
 	}
-	jd.assert(newDSIdx != "", fmt.Sprintf("newDSIdx is empty"))
+	jd.assert(newDSIdx != "", "newDSIdx is empty")
 
 	defer func() {
 		if appendLast {
@@ -1772,7 +1898,9 @@ func (jd *HandleT) getUnprocessedJobsDS(ds dataSetT, customValFilters []string,
 	}
 
 	result := hasJobs
-	if len(jobList) == 0 {
+	dsList := jd.getDSList(false)
+	//if jobsdb owner is a reader and if ds is the right most one, ignoring setting result as noJobs
+	if len(jobList) == 0 && (jd.ownerType != Read || ds.Index != dsList[len(dsList)-1].Index) {
 		jd.logger.Debugf("[getUnprocessedJobsDS] Setting empty cache for ds: %v, stateFilters: NP, customValFilters: %v, parameterFilters: %v", ds, customValFilters, parameterFilters)
 		result = noJobs
 	}
@@ -1890,6 +2018,19 @@ func (jd *HandleT) addNewDSLoop() {
 	}
 }
 
+func (jd *HandleT) refreshDSListLoop() {
+	for {
+		time.Sleep(refreshDSListLoopSleepDuration)
+		jd.logger.Debugf("[[ %s : refreshDSListLoop ]]: Start", jd.tablePrefix)
+
+		jd.dsListLock.Lock()
+		jd.getDSList(true)
+		jd.getDSRangeList(true)
+
+		jd.dsListLock.Unlock()
+	}
+}
+
 func (jd *HandleT) migrateDSLoop() {
 
 	for {
@@ -1916,7 +2057,16 @@ func (jd *HandleT) migrateDSLoop() {
 			ifMigrate, remCount := jd.checkIfMigrateDS(ds)
 			jd.logger.Debugf("[[ %s : migrateDSLoop ]]: Migrate check %v, ds: %v", jd.tablePrefix, ifMigrate, ds)
 
-			if liveDSCount >= maxMigrateOnce || liveJobCount >= maxDSSize || idx == len(dsList)-1 {
+			var idxCheck bool
+			if jd.ownerType == Read {
+				//if jobsdb owner is read, expempting the last two datasets from migration.
+				//This is done to avoid dslist conflicts between reader and writer
+				idxCheck = (idx == len(dsList)-1 || idx == len(dsList)-2)
+			} else {
+				idxCheck = (idx == len(dsList)-1)
+			}
+
+			if liveDSCount >= maxMigrateOnce || liveJobCount >= maxDSSize || idxCheck {
 				break
 			}
 
@@ -2197,7 +2347,7 @@ func (jd *HandleT) backupTable(backupDSRange dataSetRangeT, isJobStatusTable boo
 		batchCount++
 
 		//Asserting that the first character is '[' and last character is ']'
-		jd.assert(rawJSONRows[0] == byte('[') && rawJSONRows[len(rawJSONRows)-1] == byte(']'), fmt.Sprintf("json agg output is not in the expected format. Excepted format: JSON Array [{}]"))
+		jd.assert(rawJSONRows[0] == byte('[') && rawJSONRows[len(rawJSONRows)-1] == byte(']'), "json agg output is not in the expected format. Excepted format: JSON Array [{}]")
 		rawJSONRows = rawJSONRows[1 : len(rawJSONRows)-1] //stripping starting '[' and ending ']'
 		rawJSONRows = append(rawJSONRows, '\n')           //appending '\n'
 
@@ -2336,14 +2486,14 @@ func (jd *HandleT) JournalMarkStart(opType string, opPayload json.RawMessage) in
 		opType == dropDSOperation ||
 		opType == RawDataDestUploadOperation, fmt.Sprintf("opType: %s is not a supported op", opType))
 
-	sqlStatement := fmt.Sprintf(`INSERT INTO %s_journal (operation, done, operation_payload, start_time)
-                                       VALUES ($1, $2, $3, $4) RETURNING id`, jd.tablePrefix)
+	sqlStatement := fmt.Sprintf(`INSERT INTO %s_journal (operation, done, operation_payload, start_time, owner)
+                                       VALUES ($1, $2, $3, $4, $5) RETURNING id`, jd.tablePrefix)
 	stmt, err := jd.dbHandle.Prepare(sqlStatement)
 	jd.assertError(err)
 	defer stmt.Close()
 
 	var opID int64
-	err = stmt.QueryRow(opType, false, opPayload, time.Now()).Scan(&opID)
+	err = stmt.QueryRow(opType, false, opPayload, time.Now(), jd.ownerType).Scan(&opID)
 	jd.assertError(err)
 
 	return opID
@@ -2358,8 +2508,8 @@ func (jd *HandleT) JournalMarkDone(opID int64) {
 
 //JournalMarkDoneInTxn marks the end of a journal action in a transaction
 func (jd *HandleT) journalMarkDoneInTxn(txHandler transactionHandler, opID int64) error {
-	sqlStatement := fmt.Sprintf(`UPDATE %s_journal SET done=$2, end_time=$3 WHERE id=$1`, jd.tablePrefix)
-	_, err := txHandler.Exec(sqlStatement, opID, true, time.Now())
+	sqlStatement := fmt.Sprintf(`UPDATE %s_journal SET done=$2, end_time=$3 WHERE id=$1 AND owner=$4`, jd.tablePrefix)
+	_, err := txHandler.Exec(sqlStatement, opID, true, time.Now(), jd.ownerType)
 	if err != nil {
 		return err
 	}
@@ -2367,8 +2517,8 @@ func (jd *HandleT) journalMarkDoneInTxn(txHandler transactionHandler, opID int64
 }
 
 func (jd *HandleT) JournalDeleteEntry(opID int64) {
-	sqlStatement := fmt.Sprintf(`DELETE FROM %s_journal WHERE id=$1`, jd.tablePrefix)
-	_, err := jd.dbHandle.Exec(sqlStatement, opID)
+	sqlStatement := fmt.Sprintf(`DELETE FROM %s_journal WHERE id=$1 AND owner=$2`, jd.tablePrefix)
+	_, err := jd.dbHandle.Exec(sqlStatement, opID, jd.ownerType)
 	jd.assertError(err)
 }
 
@@ -2379,7 +2529,9 @@ func (jd *HandleT) GetJournalEntries(opType string) (entries []JournalEntryT) {
 									done=False
 									AND
 									operation = '%s'
-									ORDER BY id`, jd.tablePrefix, opType)
+									AND
+									owner='%s'
+									ORDER BY id`, jd.tablePrefix, opType, jd.ownerType)
 	stmt, err := jd.dbHandle.Prepare(sqlStatement)
 	jd.assertError(err)
 	defer stmt.Close()
@@ -2398,7 +2550,7 @@ func (jd *HandleT) GetJournalEntries(opType string) (entries []JournalEntryT) {
 	return
 }
 
-func (jd *HandleT) recoverFromCrash(goRoutineType string) {
+func (jd *HandleT) recoverFromCrash(owner OwnerType, goRoutineType string) {
 
 	var opTypes []string
 	switch goRoutineType {
@@ -2418,7 +2570,9 @@ func (jd *HandleT) recoverFromCrash(goRoutineType string) {
 									done=False
 									AND
 									operation = ANY($1)
-                                	ORDER BY id`, jd.tablePrefix)
+									AND
+									owner = '%s'
+                                	ORDER BY id`, jd.tablePrefix, owner)
 
 	stmt, err := jd.dbHandle.Prepare(sqlStatement)
 	jd.assertError(err)
@@ -2518,15 +2672,16 @@ const (
 	migratorRoutine = "migrator"
 )
 
-func (jd *HandleT) recoverFromJournal() {
-	jd.recoverFromCrash(addDSGoRoutine)
-	jd.recoverFromCrash(mainGoRoutine)
-	jd.recoverFromCrash(backupGoRoutine)
+func (jd *HandleT) recoverFromJournal(owner OwnerType) {
+	jd.recoverFromCrash(owner, addDSGoRoutine)
+	jd.recoverFromCrash(owner, mainGoRoutine)
+	jd.recoverFromCrash(owner, backupGoRoutine)
 }
 
 //RecoverFromMigrationJournal is an exposed function for migrator package to handle journal crashes during migration
 func (jd *HandleT) RecoverFromMigrationJournal() {
-	jd.recoverFromCrash(migratorRoutine)
+	jd.recoverFromCrash(Write, migratorRoutine)
+	jd.recoverFromCrash(ReadWrite, migratorRoutine)
 }
 
 /*
